CORE-5001 Refactor TransactionMetaData interface to be a class in runtime-os
